dev_requirements.txt
docker==6.0.1
requests==2.28.1
Faker==13.15.0
pytest==7.2.1
pytest-asyncio==0.20.3
black==24.4.2
pylint
pylint_django==2.5.5
pytest-cov==5.0.0